Moved the statements contained in the if block of the host stub (which only executed on 1st time execution of the host) into the outer scope

+++ scripts/runTests.py
@@ -79,9 +79,9 @@ def clean ():
 	filesToRemove.append('testReport.txt')
 
 	for file in filesToRemove:
-		if opts.verbose:
-			print("Removing file: '" + file + "'") 
-		os.remove(file)
+		if os.path.exists(file):
+			debug.verboseMessage("Removing file: '" + file + "'")
+			os.remove(file)
 
 def generateCUDAMakefile (generatedFiles, testNum):
 	from FileDependencies import getBaseFileName, determineModuleDependencies
